# src/products/dto.ts/product.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class ProductDto {
  @ApiProperty()
  id: string;

  @ApiProperty()
  sku: string;

  @ApiProperty()
  name: string;

  @ApiProperty()
  brand: string;

  @ApiProperty()
  model: string;

  @ApiProperty()
  category: string;

  @ApiProperty()
  color: string;

  @ApiProperty()
  price: number;

  @ApiProperty()
  currency: string;

  @ApiProperty()
  stock: number;

  @ApiProperty()
  locale: string;

  @ApiProperty()
  createdAt: string;

  @ApiProperty()
  updatedAt: string;
}